defensive-ops/kumoshield/detection/requirements.txt
# KumoShield Detection Engine - Python Dependencies

# Sigma Rules
pysigma==0.10.4
sigma-cli==0.7.8
pyyaml==6.0.1

# YARA
yara-python==4.5.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# Isolation Forest (Anomaly Detection)
pyod==1.1.2

# Performance
numba==0.58.1

# Kafka Integration
kafka-python==2.0.2
confluent-kafka==2.3.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-benchmark==4.0.0
